requirements.txt
Django~=2.2.10
django-widget-tweaks
django-allauth~=0.41.0  
Pillow
Markdown
django-mdeditor
markdownx
# pip3 install -r requirements.txt